Utility Data Fetcher CSV
- Now the Utility data is fetched directly from the repo instead of cloning the repo.
